[art] Fix app image validation of dex files already present in the CL
As part of app image validation ART will verify that the runtime
classloader matches the classloader encoded in the app image. The logic
is as follows:

- Extract dex file names from currently loading dex elements
- Extract dex file names from image classloader
- Remove the entries in the list of image classloader names that
correspond to the dex cache entries in the image file, since those
dexElements haven't been loaded into the runtime classloader yet
- Compare the two lists of dex file names

This is missing extracting dex file names from the runtime classloader,
and in the implementation details the dex file names are added forward
from the currently loading dex elements while they're added backward
from the classloaders.

That is, both extractions linearly scan the dex elements, but the list of
loading dex elements dex file names is computed by pushing back each new file
name, while the list of classloader dex file names is computed by pushing each
new file name to the front of the list.

In effect the list of image dex file names is in reverse order, while
the list of runtime dex file names is in forward order, so the
comparison will only pass when both lists should be empty.

In addition, since the list of image dex files is in reverse order, the
resize that's done incorrectly truncates the "oldest" dex files, instead
of the "newest" ones[1], thus the comparison makes even less sense.

---

Notes on implementation to fix the ordering issue:
There are a few possible implementations of the fix:

(1) We can change GetDexFileNames to push_back each new dex file name, which
appears to be safe since the only other caller is CompareClassLoaders,
and CompareClassLoaders isn't sensitive to order of dex file names
(since it uses GetDexFileNames to extract the two sets of file names its
comparing). This is the approach I opted for since the change is pretty
small then.

(2) We can change how we extract dex file names from the runtime classloader
in OpenImageSpace (i.e. reverse the order of that list) and change how
we remove the image dex file names corresponding to dex cache entries
(i.e. resize won't work here, we'll need to pop_front).

(3) We can reverse the list of image dex file names after we extract
them in OpenImageSpace just before we resize. This would be a similarly
small change (a single line).

I opted for the first solution over the last two because the last two
force the runtime to compute more data that the first (in (2) we would
need to have a while loop to pop_front image_size times and in (3) we
would be forced to additionally reverse a list).

Test: ./test/testrunner/testrunner.py -b --host
Test: Added 597-app-images-same-classloader (runs with above command)
Change-Id: If06f4c286e090569821d77d26f7fffce50473614