Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Thanks for putting together this quick fix. I think it'd be cleaner if some of the copying logic was moved to Android.mk (see comment).

This was intended as a fix just to move things forward and stop the relentless stream of build failures. Copying libraries out of the conscrypt apex seems pretty ugly whether it's in .mk or in bash, but fine.

Patch-set: 2
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0