Update patch set 3
Patch Set 3:

As-is this doesn't attempt to minimize the size of the delta here. Primary goal at the moment is to see whether this unbreaks go/lem. We can tidy if so (though it'll still be a mis-use of a build system).

Patch-set: 3
CC: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>