Update patch set 1
Patch Set 1:

I have a comment thread https://docs.google.com/document/d/1XouNOC1iCJ1RCZjhjSaqBxmrbMD6zVPxILkCmKX8hD0/edit?disco=AAAAI2Yha_s discussing this if you could chime in there.

Is it that useful?  Would you want it in logcat itself too?  It's only in the host side now and given the rest of the comments on that doc + the symmetry with the device it seemed fine to drop it.

Patch-set: 1