Update patch set 1
Patch Set 1:

> I have a comment thread https://docs.google.com/document/d/1XouNOC1iCJ1RCZjhjSaqBxmrbMD6zVPxILkCmKX8hD0/edit?disco=AAAAI2Yha_s discussing this if you could chime in there.
> 
> Is it that useful?  Would you want it in logcat itself too?  It's only in the host side now and given the rest of the comments on that doc + the symmetry with the device it seemed fine to drop it.

I suggest discussing this on the doc thread to keep it in one place.

In the meantime I'll wait a bit with reviewing this topic. I think we should also let https://r.android.com/1202478 get in and brew for a week or two to see what falls out before continuing with this.

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>