Additional work on packaging the go/lem tarball
Delete existing symlinks if they're going to be copied
over (necessary for dex2oat).

Update flags to cp to preserve relative symlink. This is necessary for
dalvikvm since ART script determines whether to use a library path for
32-bit or a 64-bit based on the target name of the dalvikvm symlink.

Add dexdump to programs copied to system/bin to trigger test run on
go/lem which does not otherwise detect this CL as changing the build.

Bug: 129332183
Test: Manual inspection of tarball for 32-bit and 64-bit
Test: Still able to run benchmarks on local 32-bit test device
Test: go/lem trial of CL

Change-Id: I24819a36bc7f0370c3778236b5e7f7308f3b48d8