Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> What's the rationale to change the current behaviour?

The high-level end goal is to enable app images for secondary dex files. Before making that change I wanted to increase test coverage for secondary app images. To this end, this CL changes the default behavior of ART tests to include secondary app images so that we horizontally improve test coverage for secondary app image. Let me know if I can provide any further clarification!

It might be a good idea to introduce a new "compiler-type" so that we have an instance of a test with and without secondary app images as opposed to enabling secondary app images wholesale across all the tests. What do you think?

Patch-set: 4