Fix Transaction constraint validation...
... for boot image extensions. Add WriteConstraint checks
to APUT instructions and add necessary WriteConstraint and
WriteValueConstraint checks to UnstartedRuntime.

For strict transactions (app compilation), prevent writing
to boot image objects. However, more work is required for
this use case as the UnstartedRuntime needs a review for
missing ReadConstraint checks and the WriteValueConstraint
may need to be more restrictive.

While the transaction_test is improved to test Transaction
constraints more thoroughly, no regression tests are
provided for the previously missing checks. Such tests are
difficult to write as they would the require compilation of
a custom boot image.

Test: Manual, include java.lang.Locale[] in primary
      boot image (hack CompilerDriver) and build.
      This previously aborted in ImageWriter.
Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 119800099
Change-Id: Ibfe1b24b10dbd982b4e4ae4d98289e587a842812