Allow removal of unused, side-effect-free LoadClass instructions
We accomplish this through several steps.

First, we add a bit to the HLoadClass instruction that tracks whether the class
is known to already be initialized at the point that instruction executes. This
essentially stores the value of IsInitialized() in instruction_builder.cc.

Second, we tweak the logic in IsInitialized. Instead of doing is_subclass &&
HasTrivialInitialization, we switch this to ||. In the subclass case, we know
the class has already been initialized due to the control flow. On the other
hand, if we have trivial initialization, there is nothing that needs to be done.

Using these two changes, we can then make HLoadClass::CanThrow more precise. If
the class is known already to be initialized, then the instruction will not be
able to throw since the class is already loaded. This makes the HLoadClass
instruction susceptible to dead code elimination later if all of its uses are
removed.

Bug: 147839012
Change-Id: I54008c02736a340793da40ff1ca29bea4e869d07