Update patch set 6
Patch Set 6:

Meta-level comment:
I view this CL as a workaround that we have to do on the critical path of the product (stack walking), because a debugging feature (instrumentation) is not implemented right. Without external constraints, this is something we should avoid.

I also think we should have the time to tweak the debugging feature to end up in a proper state. With the observation that:
1) The current issue being resolved is not a regression compared to old releases.
2) Working around the issue on the main path can have unintended consequences on the quality of the product
3) A brief look at the instrumentation code makes me think it shouldn't take too long to do it right.

Patch-set: 6
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>