Update patch set 3
Patch Set 3:

> > I've also seen cases where earlier LSE makes the loop optimizations work better. It seems that the loop optimizer is pretty conservative in the presence of memory operations, so if we can remove some of them, the loop optimizer is actually able to eliminate some loops it couldn't before.
> 
> It might make sense to add another instance of LSE pass (compile-time overhead to be evaluated) instead of moving it earlier in the pipeline: other optimizations (GVN, DCE and LICM) might expose some opportunities for LSE which would have been impossible otherwise.

Ah, good idea. I suspect the limit of this line of inquiry is to run a set of passes in a loop for a few iterations, since each one tends to make the others work better. Of course, we'll have to keep an eye on the compile time impact.

Patch-set: 3
Reviewer: Gerrit User 1335232 <1335232@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0