Update patch set 2
Patch Set 2:

I'm not sure that this isn't simply replacing one undefined behavior with another?

With this change if we had something like a 16-bit nested within a 32-bit it would look like

  union {
    struct { int16; } /* e.g. offset 0, width 16 */
    struct { int32; } /* e.g. offset 16, width 16 ... OR offset 0 width 32 */
  }

This is because StorageType is expanded to be the smallest type necessary to hold bitWidth+bitOffset.
-----

However, IIUC, it is undefined behavior to inspect a member of a union (other than the one most recently written to) if it is not part of the common initial subsequence.

Two types can be part of the common initial subsequence if they are layout-compatible. For trivial types that means they have to be the same type ignoring cv qualifiers.

In the example above int16 != int32, so this seems like it's not fitting the definition for layout-compatible, therefore its UB?

Recommend this as a starting point of discussion: https://en.cppreference.com/w/cpp/language/data_members#Standard_layout

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>