Update patch set 2
Patch Set 2:

(3 comments)

It seems to be the original code (and this CL) was possibly incorrectly interpreting the definition of 'layout-compatible' ?

https://timsong-cpp.github.io/cppwp/n4659/basic.types#11

> Two types cv1 T1 and cv2 T2 are layout-compatible types if T1 and T2 are the same type, layout-compatible enumerations, or layout-compatible standard-layout class types.

A struct's common initial sequence is the longest sequence of its members that have layout-compatible types.

https://timsong-cpp.github.io/cppwp/n4659/class.mem#20

---

C++20 adds an is_layout_compatible type trait to further demystify this. Sadly neither trunk gcc or trunk clang (at least on godbolt) appear to have implemented this yet to do a quick test.

---

The fundamental assumption here made in BitStructField is the same as in thread.h StateAndFlags, i.e. that

   union {
      struct X {
        int16;
        int16;
      }
      int32 Y;
   }

Old Assumption: 'X' and 'Y' share a common initial sequence and are layout-compatible.

But given the above interpretation of the definitions I think this may also be incorrect, two structs are not layout compatible even if their sizeof is the same.

----------

Personal note:

It seems to me this definition of 'layout-compatible' is overly strict, as it is often possible to write structs where the offsetof each member is identical without each type being the same.

I suspect they may have wanted to preserve strong type aliasing (ignoring cv-qualifiers).

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>