Update patch set 2
Patch Set 2:

(10 comments)

I don't really like the way this is using a std::map as a stack (using the defined iteration order to do so). It's just asking for someone to come around with the bright idea to change it to unordered_map for efficiency and break everything.

Patch-set: 2
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>