Remove build target `test-art-target-sync`.
This build target is essentially dead code, as we have been relying on
the `tools/buildbot-sync.sh` script to sync build products to the
device for some time now.

Note that this build target has been broken since Android builds
became sandboxed and disallowed the execution of `adb`.

Also remove build target `test-art-target-run-test-dependencies`, as
it is just a dependency on `test-art-target-sync`.

Test: Rely on the Android Build and the ART Buildbot
Bug: 111676396
Change-Id: Ib28aa6c24eed892605c7bac1bccbcbe8d72ac12b