Device gtests: Use boot.art instead of core.art.
They are essentially the same. We can use boot.art to run gtests
since it is already part of the apex, including the jar files.

This will make it easier to run the tests in atest, since we
will not have to worry about copying core.art to the device.

The long-term goal is to avoid generating core.art altogether.

Couple of tests also require "alternate" image which has no
compiled code. The tests now generate it on-demand in code.

The host gtests still use core.art for now (as there is no
boot.art on host). The plan is to add it in future CLs.

Test: m test-art-host-gtest
Test: ./art/tools/run-gtests.sh
Bug: 147817606
Change-Id: I3a750bb8e60eea0b4a7df1491285feffd5a0161c