Generate a linker configuration file for chroot-based ART testing.
Build `/system/bin/linkerconfig` as part of ART chroot-based target
testing builds, include it in sync'd files, and invoke it in
`tools/buildbot-sync.sh` to generate a linker configuration file for
the chroot environment. Adjust the linker configuration file path
logic in `tools/buildbot-sync.sh` to match the one in top-of-tree
Bionic linker code.

Test: Run ART tests in a chroot environment on an Android R "host device"
Test: Run ART tests in a chroot environment on an Android Q "host device"
Bug: 147737840
Bug: 148171362
Change-Id: I457d8c75988b6a9254b7ac9db3642f40f4271f1d