Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > runtime_callbacks_test will fail. Checking it.
> 
> When running ThreadLifecycleCallbackRuntimeCallbacksTest, it failed at below line (No pending exception expected: java.lang.NullPointerException).
> 
> env->CallVoidMethod(thread.get(), start_id);
> 
> 
> From callstack, the error occurs when trying to invoking the run function of java.lang.Thread (InvokeVirtualOrInterfaceWithJValues).

I have identified that the exception comes from the State.values() in getState() of java.lang.Thread (I'am sure it is not related to nativeGetStatus() in it). In my device, the priority and sched policy of a thread is set correctly with this patch, so I think the getState() should be fine in real case, but I don't understand why it fails when running RuntimeCallbacksTest of art-test. The test will pass if I modify getState() as below, but I'm not sure if it is ok to do so.

Tobias/Orion,

Could you please provide some advice. Thanks. :)

--- a/ojluni/src/main/java/java/lang/Thread.java
+++ b/ojluni/src/main/java/java/lang/Thread.java
@@ -2016,7 +2016,14 @@ class Thread implements Runnable {
         // Use Android specific nativeGetStatus() method. See comment on started field for more
         // information.
         // return sun.misc.VM.toThreadState(threadStatus);
-        return State.values()[nativeGetStatus(started)];
+        State state;
+        try {
+            state = State.values()[nativeGetStatus(started)];
+        } catch (Exception e) {
+            state = State.TERMINATED;
+        }
+
+        return state;
     }

Patch-set: 2
Reviewer: Gerrit User 1324603 <1324603@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0