Dump instruction set features in .cfg
This commit adds a compilation block at the beginning of the .cfg
dumped by the optimizing compiler when --dump-cfg is enabled.

The compilation block appears in the following form:

begin_compilation
  name "isa_features:feature1,-feature2"
  method "isa_features:feature1,-feature2"
  date 1580721972
end_compilation

This compilation block dump is passed to checker script
https://android-review.googlesource.com/c/platform/art/+/1274876
for checking if a certain instruction set feature was used at compile
time.

Test: checker/run_unit_tests.py
Test: test.py --target --optimizing
Change-Id: I6dc658a9d79f76c6929d5d6900aa5fc56724f07c