Update patch set 1
Patch Set 1:

I think we're missing a check on the method/field flags that the targets are actually part of the core platform :-). That's probably 5 lines of code to fix, but I can't see doing it remotely this week so please revert.

Patch-set: 1
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>