Update patch set 1
Patch Set 1:

This has been in limbo for a long time. Concerns have been discussed elsewhere, but I'll document them here briefly for the record. 

My major concern is that it's unclear from the architecture spec whether lock addl $0,0(SP) is safe if another thread is concurrently accessing 0(SP) with a different access granularity, which doesn't appear to be precluded.

A secondary concern, which is perhaps not a showstopper given the performance difference, is that, under unlikely conditions, this can cause cache contention (coherence misses), which would be very unexpected and potentially hard to diagnose.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>