Update patch set 5
Patch Set 5:

I'm experimenting with a similar CL and a modified version of 2029-contended-monitors as a microbenchmark. I think we actually want to be quite aggressive about NOT inflating due to contention. Once we inflate, it's hard to go back, and we still seem to lose 1+ orders of magnitude in performance in the fat locked case, if a thin lock would have done. Based on the microbenchmark, spinning for 16K iterations, with a very small number (8) of sched_yields thrown in now and then, seems to get good performance, with a similar delay before inflation as the old code. Pure spinning, as suggested here, does seem to be much more effective per unit time.

Patch-set: 5
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>