Speed up test
Reduce the size of the arrays to make this cheaper.

We expect that especially with gcstress this used to accidentally
be much cheaper. Since the main thread exits immediately, we started
the shutdown process before most of the work was done, and thus
this was inadvertently largely run with the GC disabled.

Since it is now WAI, it took too long, and timed out regularly.

Bug: 148126377
Test: TreeHugger
Change-Id: I42e4e3ff04d261f6beaaa942e178a78a91ee5f78