Update patch set 1
Patch Set 1:

(4 comments)

Every single one of the dependencies exported here is needed to build conscrypt. Exactly why I don't know and I don't have either the knowledge or time to investigate and remove unnecessary dependencies. So, simply saying that they *shouldn't* be exported is not particular helpful as at the moment they *need* to be exported to meet the goals of building conscrypt unbundled.

So, where you have concerns could you suggest one of the following actions and I'll fix it:
1) If it should just not be part of the sdk (because it is not provided by the apex) then I'll move it to the host (probably badly named now) module_exports.
2) If I have the contents of the conscrypt/art modules incorrect let me know then I can update my tests and update this accordingly.
3) Otherwise, raise a bug to fix whatever issue and suggest a comment for me to add to the reference in the sdk/module_exports. e.g. but not limited to:
a) If usages outside ART are invalid and need removal.
b) If it needs stubs.
c) If it is meant to just be an implementation detail of an export. This would need extra support in the sdk code which I will be happy to do.

This is not intended to be 100% perfect or a reflection what *should* be. It is, no pun intended, simply a snapshot of what currently *is*. It is intended to provide information that will allow us to identify issues so they can be resolved in future.

Apologies if this comes across as acerbic (or even rude) that is not intentional. It has just been a quite frustrating process to get this far. 😂

Patch-set: 1
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0