Update patch set 1
Patch Set 1:

(1 comment)

The need for these in the SDK most likely means conscrypt isn't built correctly. We have to figure out the reasons for that and fix them.

If this helps unblock some work for you then I've got no problem with it going in as a temporary measure, but then there should be TODO's in the bp file next to the incorrect exports to make that clear, and a ticket to track. For the ART native side I'm happy to own the ticket initially.

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>